A lot of shit happened...
