--- WillingHuskyModularity/README.md
@@ -1,2 +1,0 @@
-# WillingHuskyModularity
-Prototype for dictionary based menus in python
