--- WillingHuskyModularity/main.py
@@ -1,4 +1,0 @@
-from menu_parse import traverse_menu
-from menu import proto
-
-traverse_menu(proto)
