+++ control-scripts/README.md
@@ -1,1 +1,14 @@
 # control-scripts
+## ledctrl
+Simple script to control the state of the fitstatusb stick.
+
+Usage: led_controller.py <option>
+
+Options:
+  -h,--help 	Print this usage info and exit.
+  -a,--attn 	Set the LED state to attention mode.
+  -b,--blink 	Set the LED state to blinking mode.
+  -d,--default 	Reset the LED state to deafult mode.
+  -e,--error 	Set the LED state to error mode.
+  -v,--version 	Print the version number and exit.
+
