+++ control-scripts/install.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# create symbolic links for the control scripts in /usr/local/bin
+ln ledctrl.py /usr/local/bin/ledctrl
