--- fitStatUSB-scripts/fitStatUSB-setup.service
@@ -1,14 +1,0 @@
-[Unit]
-Description=Run initialization for using fitStatUSB device.
-After=network.target
-
-[Service]
-ExecStart=/bin/sh -e fitStatUSB-setup.sh
-WorkingDirectory=/home/pi/pimera-cluster/fitStatUSB-scripts
-StandardOutput=inherit
-StandardError=inherit
-Restart=always
-User=pi
-
-[Install]
-WantedBy=multi-user.target
