--- led-controller/README.md
@@ -1,2 +1,0 @@
-# led-controller
-daemon service to handle/control the state of the fitstatusb stick.
