+++ services/fsusb/tty_device
@@ -1,0 +1,1 @@
+/dev/ttyACM0
